python-process/visualization-task/requirements.txt
rerun-sdk==0.23.3
boto3~=1.38.22
opencv-python-headless
cantools~=40.2.1
av~=14.4.0
numpy~=2.2.1
paho-mqtt~=2.1.0
numba~=0.61.2
psutil

